Simplified interaction files organisation | Encountered massive problem regarding interaction permissions
