+++ main.ts
@@ -1,9 +1,9 @@
 // Lib
-import { ComponentInteractionHandler, CommandClient, Interaction } from "./deps.ts";
+import { CommandClient } from "./deps.ts";
 // Src
 import { ready, loadConfig, msgCreate, GuildSettingsService, initTemp, initLocalDb, Command, commandList } from './src/mod.ts'
 // Slash commands
-import { Interactions, InteractionHandlers} from './src/interactions/interactions.ts'
+import { Interactions, InteractionHandlers, ComponentInteractionHandlers } from './src/interactions/interactions.ts'
 
 // --- BEGIN ---
 
@@ -51,30 +51,26 @@ Interactions.forEach(command => {
 	// If you want to create command globally, just remove 'Your Server/Guild ID' part
 	// I recommend making it for only one guild for now because Global Slash Commands can take max 1 hour to come live.
 	BOT.interactions.commands
+		// TODO : see if possible with Harmony to require server wide permission for slash commands
+		// else fork harmony and implement it
 		.create(command, '376040838540820481')
-		.then((cmd) => console.log(`Created Slash Command ${cmd.name}!`))
+		// TODO : register in the DB cmd.id & cmd.guid to delete all commands when needed
+		.then((cmd) => {
+			console.log(`Created Slash Command ${cmd.name}!`)
+			//cmd.setPermissions([{id: '376040838540820481', type: "", permission: true}])
+		})
 		.catch((e) => console.log(`Failed to create command : ${command.name}`));
 })
-
-BOT.interactions.commands.delete('1341157964484055120', '376040838540820481')
 
 // Push all interaction handlers
 InteractionHandlers.forEach(handler => {
 	BOT.interactions.handlers.push(handler)
 })
 
-const ketHandler:ComponentInteractionHandler = {
-	customID: '112229907884085659',
-	type: 'button',
-	handler: (i: Interaction)=> {
-		if (i.message){
-			i.message.delete()
-		}
-		i.message?.channel.send('Good job, here\'s your ketamine \n (*`2g`* of ketamine added to your inventory)')
-	}
-} 
-BOT.interactions.componentHandlers.push(ketHandler)
-
+// Push all component interaction handlers
+ComponentInteractionHandlers.forEach(handler => {
+	BOT.interactions.componentHandlers.push(handler)
+})
 
 // --- Launch ---
 BOT.connect()
