+++ src/interactions/ClearMessages/clearMsgs.ts
@@ -1,10 +1,13 @@
 import { ApplicationCommandHandler, ApplicationCommandPartial, ApplicationCommandType, SlashCommandOptionType } from "../../../deps.ts";
 import { ClearMsgs } from "./clearMsgs.function.ts";
 
+
+
 // Creater interation object
 export const ClearMsgsInteraction: ApplicationCommandPartial = {
 	name: 'clearmessages',
 	description: 'Deletes a bunch of messages in current channel',
+	defaultPermission: false,
 	options: [
 		{
 			name: 'count',
@@ -17,7 +20,7 @@ export const ClearMsgsInteraction: ApplicationCommandPartial = {
 
 // Create interaction handler
 export const ClearMsgsHandler: ApplicationCommandHandler = {
-	name: 'clearMsgs',
+	name: 'clearmessages',
 	type: ApplicationCommandType.CHAT_INPUT,
 	handler: ClearMsgs
 }
