+++ src/mod.ts
@@ -5,15 +5,11 @@ export * from './core/format.ts'
 export * from './core/config.ts'
 
 // Types
-export * from './types/commandTypes/horoscope.ts'
-export * from './types/commandTypes/downloadEmotes.ts'
 export * from './types/core.ts'
 export * from './types/consts.ts'
 export * from './types/enums.ts'
 
 // Helpers
-export * from './helpers/commandHelpers/downloadEmotes.ts'
-export * from './helpers/commandHelpers/horoscope.ts'
 export * from './helpers/commandHelpers/random.ts'
 export * from './helpers/commandHelpers/triggers.ts'
 export * from './helpers/dateTime.ts'
@@ -30,4 +26,4 @@ export * from './handlers/ready.ts'
 export * from './handlers/msg.create.ts'
 
 // Commands
-export * as commandList from './commands/commandList.ts'
+export * as commandList from './commands/commands.ts'
