--- src/types/commandTypes/downloadEmotes.ts
@@ -1,7 +1,0 @@
-export type Emote = {
-	animated: boolean
-	id: string
-	name: string
-	filename: string
-	url: string
-}
